refactor: 下沉原生 Agent 协议与模型
