+++ packages/agentcore/lib/agentcore.dart
@@ -5,3 +5,6 @@ export 'src/contracts.dart';
 export 'src/json_turn_parser.dart';
 export 'src/memory/agent_memory.dart';
 export 'src/permission/agent_authorization.dart';
+export 'src/model/native_tool_agent_model.dart';
+export 'src/protocol/native_tool_protocol.dart';
+export 'src/protocol/openai_compatible_transport.dart';
